FHAC-201:Fix and cleanup based on PR comments
